Refactored frontend startup into builder pattern.
It is now also possible to listen on plain http again.
