+++ api/proxy.go
@@ -21,15 +21,11 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"errors"
-	"fmt"
 	"net/http"
 	"net/http/httputil"
 	"net/url"
-	"sync"
-	"time"
 
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
-	"github.com/sirupsen/logrus"
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
@@ -92,7 +88,7 @@ func AddProxyMux(config_obj *config_proto.Config, mux *http.ServeMux) error {
 }
 
 // Prepares a mux by adding handler required for the GUI.
-func PrepareMux(config_obj *config_proto.Config, mux *http.ServeMux) (http.Handler, error) {
+func PrepareGUIMux(config_obj *config_proto.Config, mux *http.ServeMux) (http.Handler, error) {
 	ctx := context.Background()
 	h, err := GetAPIHandler(ctx, config_obj)
 	if err != nil {
@@ -161,83 +157,6 @@ func PrepareMux(config_obj *config_proto.Config, mux *http.ServeMux) (http.Handl
 
 	err = MaybeAddOAuthHandlers(config_obj, mux)
 	return mux, err
-}
-
-func StartSelfSignedHTTPSProxy(
-	ctx context.Context,
-	wg *sync.WaitGroup,
-	config_obj *config_proto.Config, mux http.Handler) {
-	logger := logging.Manager.GetLogger(config_obj, &logging.GUIComponent)
-
-	cert, err := tls.X509KeyPair(
-		[]byte(config_obj.Frontend.Certificate),
-		[]byte(config_obj.Frontend.PrivateKey))
-	if err != nil {
-		logger.Error("GUI Error", err)
-		return
-	}
-
-	listenAddr := fmt.Sprintf("%s:%d",
-		config_obj.GUI.BindAddress,
-		config_obj.GUI.BindPort)
-
-	server := &http.Server{
-		Addr:     listenAddr,
-		Handler:  mux,
-		ErrorLog: logging.NewPlainLogger(config_obj, &logging.FrontendComponent),
-
-		// https://blog.cloudflare.com/the-complete-guide-to-golang-net-http-timeouts/
-		ReadTimeout:  500 * time.Second,
-		WriteTimeout: 900 * time.Second,
-		IdleTimeout:  15 * time.Second,
-		TLSConfig: &tls.Config{
-			MinVersion:               tls.VersionTLS12,
-			CurvePreferences:         []tls.CurveID{tls.CurveP521, tls.CurveP384, tls.CurveP256},
-			Certificates:             []tls.Certificate{cert},
-			PreferServerCipherSuites: true,
-			CipherSuites: []uint16{
-				tls.TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,
-				tls.TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
-				tls.TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA,
-				tls.TLS_RSA_WITH_AES_256_GCM_SHA384,
-				tls.TLS_RSA_WITH_AES_256_CBC_SHA,
-				tls.TLS_RSA_WITH_AES_128_GCM_SHA256,
-			},
-		},
-	}
-
-	logger.WithFields(
-		logrus.Fields{
-			"listenAddr": listenAddr,
-		}).Info("GUI is ready to handle TLS requests")
-
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-
-		err := server.ListenAndServeTLS("", "")
-		if err != nil && err != http.ErrServerClosed {
-			logger.Error("GUI Server error", err)
-		}
-	}()
-
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		<-ctx.Done()
-
-		logger.Info("Stopping GUI Server")
-		timeout_ctx, cancel := context.WithTimeout(
-			context.Background(), 10*time.Second)
-		defer cancel()
-
-		server.SetKeepAlivesEnabled(false)
-		err := server.Shutdown(timeout_ctx)
-		if err != nil {
-			logger.Error("GUI shutdown error ", err)
-		}
-		logger.Info("Shutdown GUI")
-	}()
 }
 
 type _templateArgs struct {
