--- artifacts/definitions/Generic/Utils/FetchBinaryPacked.yaml
@@ -1,23 +1,0 @@
-name: Generic.Utils.FetchBinaryPacked
-description: |
-  A version of Generic.Utils.FetchBinary which looks at packed
-  payloads.
-
-parameters:
-  - name: ToolName
-    default: Autorun
-  - name: SleepDuration
-
-sources:
-  - query: |
-      LET temp_binary <= tempfile(extension=".exe", permissions="x")
-
-      LET matching_tools = SELECT ToolName AS ArchiveTool, Filename
-      FROM parse_csv(filename="/inventory.csv", accessor="me")
-
-      SELECT * FROM foreach(row=matching_tools, query={
-        SELECT copy(filename=url(path="/", fragment=Filename, scheme="me").String,
-                    accessor="zip", dest=temp_binary) AS FullPath
-        FROM scope()
-        WHERE ToolName = ArchiveTool
-      })
